+++ src/components/BlubberFormFactory.js
@@ -1,8 +1,6 @@
 import Vue from 'vue';
 import VueFormGenerator from 'vue-form-generator';
 import VueFormWizard from 'vue-form-wizard';
-import StringHelper from './lib/StringHelper';
-import FormFactory from './lib/blubberFormGenerator/FormSchemaFactory';
 import 'vue-form-wizard/dist/vue-form-wizard.min.css';
 import BlubberFormSchemaConstructor from './lib/blubberFormGenerator/FormSchemaFactory';
 
@@ -13,69 +11,89 @@ const BlubberFormFactory = {
 	methods: {
 		buildBlubberForm: function ( createElement, Form, LabelGenerator )
 		{
-			let Tab, VGenerator, Description;
+			let Tab, VGenerator, Description, Step, Index;
 			const Tabs = [];
 			const FormSchema = new BlubberFormSchemaConstructor( Form, this, LabelGenerator );
 			FormSchema.build();
-			console.log( FormSchema.Form );
 
-			if( 0 === FormSchema.Steps.length )
+			if ( false === Array.isArray( FormSchema.Form.Steps ) )
 			{
+				Description = createElement(
+					'div',
+					{
+						attr: FormSchema.Form.Steps.description.attr,
+						domProps: FormSchema.Form.Steps.description.domProps
+					}
+				);
 
+				VGenerator = createElement(
+					'vue-form-generator',
+					{
+						props: FormSchema.Form.Steps.inner,
+						attr: FormSchema.Form.Steps.attr
+					}
+				);
+				return createElement(
+					'form-wizard',
+					{
+						attrs: FormSchema.Form.FormAttributes,
+						props: FormSchema.Form.FormProperties,
+						on: FormSchema.Form.FormEvents
+					},
+					[ Description, VGenerator ]
+				);
 			}
 			else
 			{
+				for ( Index in FormSchema.Form.Steps )
+				{
+					Step = FormSchema.Form.Steps[ Index ];
+					VGenerator = createElement(
+						'vue-form-generator',
+						{
+							props: Step.inner
+						}
+					);
 
+					Description = createElement(
+						'div',
+						{
+							attr: Step.description.attr,
+							domProps: Step.description.domProps
+						}
+					);
+
+					Tab = createElement(
+						'tab-content',
+						{
+							attr: Step.attr,
+							props: Step.tab
+						},
+						[ Description, VGenerator ]
+					);
+
+					Tabs.push( Tab );
+				}
+
+				return createElement(
+					'form-wizard',
+					{
+						attrs: FormSchema.Form.FormAttributes,
+						props: FormSchema.Form.FormProperties,
+						on: FormSchema.Form.FormEvents
+					},
+					Tabs
+				);
 			}
 			/*
 
-            this.$data.blubberFormSchema[ FormId ] = [];
-            this.$data.blubberModel[ FormId ] = {};
-            this.$data.currentFormId = FormId;
+			this.$data.blubberFormSchema[ FormId ] = [];
+			this.$data.blubberModel[ FormId ] = {};
+			this.$data.currentFormId = FormId;
 
-            VGenerator = createElement(
-                'vue-form-generator',
-                {
-                    props: {
-                        model: Generated.NodeSchema.model,
-                        schema: Generated.NodeSchema.schema,
-                        isNewModel: Generated.NodeSchema.isNewModel,
-                        multiple: Generated.NodeSchema.multiple,
-                        ref: Generated.NodeSchema.ref,
-                        tag: Generated.NodeSchema.tag
-                    }
-                }
-            );
+		}
 
-            Description = createElement(
-                'div',
-                {
-                    attr: Generated.NodeSchema.description.attr,
-                    domProps: Generated.NodeSchema.description.domProps
-                }
-            );
-
-            Tab = createElement(
-                'tab-content',
-                {
-                    attr: {
-                        id: Generated.NodeSchema.id
-                    },
-                    props: {
-                        title: Generated.NodeSchema.title,
-                        icon: Generated.NodeSchema.icon
-                    }
-                },
-                [ Description, VGenerator ]
-            );
-
-            Tabs.push( Tab );
-        }
-
-        return createElement( 'form-wizard', {
-            attrs: FormAttributes,
-            props: FormProperties
-        }, Tabs );*/
+		r*/
 		}
 	},
 	data: function ()
